+++ sanctum-token-ratio/src/u64_bps_fee/mod.rs
@@ -7,3 +7,5 @@ mod floor;
 pub use bps_fee::*;
 pub use ceil::*;
 pub use floor::*;
+
+pub const BPS_DENOMINATOR: u16 = 10_000;
